+++ backend/src/services/request/request.services.ts
@@ -9,7 +9,7 @@ export const requestService = {
       const userId: string = result.fortuneTellerId
       const notificationType = status === "ACCEPTED" ? "VERIFICATION" : "CANCELED_VERIFICATION"
       await notificationRepository.createNotification(userId, notificationType)
-      return { success: true, messege: "request status has been changed" }
+      return { success: true, message: "request status has been changed" }
     }
     return { success: false, message: "error to change request status" }
   },
